third_party/WebKit/LayoutTests/editing/selection/modify_extend/extend-by-character-005.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<body>
<div contenteditable id="root" class="editing">
<span id="span"><i>F    and seven</i> years <b>  as </b>our fathers  f upon this
continent, a new nation, conceived &nbsp;&nbsp; in Liberty,                 and dedicated to the
proposition that all 
<br>men are created equal.</span>
</div>
<div id="log"></div>
<script>
test(function () {
  var selection = getSelection();
  selection.collapse(span.firstChild.firstChild, 0);
  for (i = 0; i < 159; i++) {
    selection.modify("extend", "forward", "character");
  }
  for (i = 0; i < 158; i++) {
    selection.modify("extend", "backward", "character");
  }
  assert_equals(selection.anchorNode, span.firstChild.firstChild);
  assert_equals(selection.anchorOffset, 0);
  assert_equals(selection.focusNode, span.firstChild.firstChild);
  assert_equals(selection.focusOffset, 1);
});
</script>

</body>
